refactor(bot): move automod and spam handlers into automod module
Move autoModHandler.ts and spamHandler.ts from handlers/message/ to
functions/automod/handlers/ to establish proper module boundaries,
matching the pattern used by other function modules (general,
moderation, management).

Update all import paths in:
- handlers/messageHandler.ts (direct imports)
- handlers/message/index.ts (barrel re-exports)
- handlers/__tests__/messageHandler.test.ts (jest mock paths)

Rename test files to .spec.ts suffix per module convention.
No behavior changes; all 3276 tests pass (21 automod-related tests pass).

Fixes #1971
