+++ packages/bot/src/functions/automod/handlers/spamHandler.spec.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it, jest } from '@jest/globals'
 import type { Message } from 'discord.js'
-import { spamHandler } from '../spamHandler'
-import type { MessageContext } from '../types'
+import { spamHandler } from './spamHandler'
+import type { MessageContext } from '../../../handlers/message/types'
 
 jest.mock('@lucky/shared/services', () => ({
     autoModService: {
