+++ packages/bot/src/functions/automod/handlers/spamHandler.ts
@@ -5,7 +5,7 @@ import type {
     MessageContext,
     MessageHandler,
     MessageHandlerResult,
-} from './types'
+} from '../../../handlers/message/types'
 
 export const spamHandler: MessageHandler = {
     name: 'Spam',
