+++ packages/bot/src/handlers/messageHandler.ts
@@ -2,8 +2,8 @@ import { Events, type Client, type Message } from 'discord.js'
 import { featureToggleService } from '@lucky/shared/services'
 import { errorLog } from '@lucky/shared/utils'
 import { MessagePipeline } from './message/pipeline'
-import { autoModHandler } from './message/autoModHandler'
-import { spamHandler } from './message/spamHandler'
+import { autoModHandler } from '../functions/automod/handlers/autoModHandler'
+import { spamHandler } from '../functions/automod/handlers/spamHandler'
 import { customCommandHandler } from './message/customCommandHandler'
 import { afkHandler } from './message/afkHandler'
 import { xpHandler } from './message/xpHandler'
